🧰🎨: move remainder of text search functionality into `IDE` package
